requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
watchfiles>=0.21.0
websockets>=12.0
python-multipart>=0.0.6
pywebview>=5.1.0

# CRM Database (Phase 1)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Terrain & RF Propagation (Phase 1+)
numpy>=1.24.0
Pillow>=10.0.0
